Refactor + add/update comment metrics

+++ scripts/analyzers/comments/helper_scripts/parse_classify_single_line.php
@@ -3,25 +3,55 @@
 require('find_multiple_strings.php');
 
 if ($argc > 1) {
+
+    // Controleer of de --train optie aan het einde van de argumentenlijst staat
+    if (end($argv) == "--train") {
+        array_pop($argv);  // Verwijder de --train optie uit de argumentenlijst
+        $train = true;
+    } else {
+        $train = false;
+    }
+
     $tempFile = $argv[1];
-    $content = file_get_contents($tempFile);
 
-    $lines = explode("\n", $content);
+    if (file_exists($tempFile)) {
 
-    $classifiedLines = [];
+        $content = file_get_contents($tempFile);
 
-    foreach ($lines as $line) {
+        $lines = explode("\n", $content);
 
-        $classification = 'RELEVANT';
-        if (strposa(strtoupper($line), ['LICENSE', 'COPYRIGHT', 'AUTHOR'])) {
-            $classification = 'COPYRIGHT';
+        $classifiedLines = [];
+
+        foreach ($lines as $line) {
+
+
+            if (!strpos($line, ',""')) {
+
+
+                if ($train) {
+
+                    $classification = 'RELEVANT';
+                    if (strposa(strtoupper($line), ['LICENSE', 'COPYRIGHT', 'AUTHOR'])) {
+                        $classification = 'COPYRIGHT';
+                    }
+
+                    $classifiedLines[] = $line . $classification;
+
+                }
+                else {
+                    $classifiedLines[] = $line;
+                }
+
+
+            }
+
+
         }
-        $classifiedLines[] = $line . $classification;
 
+        if (count($classifiedLines)) {
+            file_put_contents($tempFile, implode("\n", $classifiedLines) . "\n");
+        }
 
     }
 
-    if (count($classifiedLines)) {
-       file_put_contents($tempFile, implode("\n", $classifiedLines)."\n");
-    }
 }
